+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 1.1.1 Unreleased
+## 1.1.1
 
 - Fix possible data race in repository lookup table, #58
 - Fix issue where "Source" wasn't clickable in the built-in frontend, #66